requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
python-multipart==0.0.6
pypdf==4.0.1
pandas==2.2.0
chardet==5.2.0
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
sqlite3-api==2.0.1
pydantic==2.3.0
numpy==1.26.2
scikit-learn==1.3.2
typing-extensions>=4.8.0
aiosmtplib==2.0.2
email-validator==2.1.0.post1
pytest
# LangChain and dependencies
langchain-community>=0.0.13
langchain-core>=0.1.5
langgraph>=0.0.17
pydantic-core>=2.14.0

# Ensure compatibility with Pydantic v1 usage
pydantic-settings>=2.0.0

aiofiles==23.2.1
langchain==0.0.267 